Replace inline launchQuestionnaire extension functions
Since they're not getting tracked in our codcov
https://github.com/jacoco/jacoco/issues/654
